refactor handler
